+++ src/ai/backend/manager/cli/image_impl.py
@@ -28,7 +28,7 @@ log = BraceStyleAdapter(logging.getLogger(__spec__.name))
 async def list_images(cli_ctx, short, installed_only):
     # Connect to postgreSQL DB
     async with (
-        connect_database(cli_ctx.local_config) as db,
+        connect_database(cli_ctx.local_config.db) as db,
         db.begin_readonly_session() as session,
         redis_ctx(cli_ctx) as redis_conn_set,
     ):
@@ -89,7 +89,7 @@ async def list_images(cli_ctx, short, installed_only):
 
 async def inspect_image(cli_ctx, canonical_or_alias, architecture):
     async with (
-        connect_database(cli_ctx.local_config) as db,
+        connect_database(cli_ctx.local_config.db) as db,
         db.begin_readonly_session() as session,
     ):
         try:
@@ -109,7 +109,7 @@ async def inspect_image(cli_ctx, canonical_or_alias, architecture):
 
 async def forget_image(cli_ctx, canonical_or_alias, architecture):
     async with (
-        connect_database(cli_ctx.local_config) as db,
+        connect_database(cli_ctx.local_config.db) as db,
         db.begin_session() as session,
     ):
         try:
@@ -129,7 +129,7 @@ async def forget_image(cli_ctx, canonical_or_alias, architecture):
 
 async def purge_image(cli_ctx, canonical_or_alias, architecture):
     async with (
-        connect_database(cli_ctx.local_config) as db,
+        connect_database(cli_ctx.local_config.db) as db,
         db.begin_session() as session,
     ):
         try:
@@ -156,7 +156,7 @@ async def set_image_resource_limit(
     architecture,
 ):
     async with (
-        connect_database(cli_ctx.local_config) as db,
+        connect_database(cli_ctx.local_config.db) as db,
         db.begin_session() as session,
     ):
         try:
@@ -180,7 +180,7 @@ async def rescan_images(
     if not registry_or_image:
         raise click.BadArgumentUsage("Please specify a valid registry or full image name.")
     async with (
-        connect_database(cli_ctx.local_config) as db,
+        connect_database(cli_ctx.local_config.db) as db,
     ):
         try:
             result = await rescan_images_func(db, registry_or_image, project)
@@ -192,7 +192,7 @@ async def rescan_images(
 
 async def alias(cli_ctx, alias, target, architecture):
     async with (
-        connect_database(cli_ctx.local_config) as db,
+        connect_database(cli_ctx.local_config.db) as db,
         db.begin_session() as session,
     ):
         try:
@@ -211,7 +211,7 @@ async def alias(cli_ctx, alias, target, architecture):
 
 async def dealias(cli_ctx, alias):
     async with (
-        connect_database(cli_ctx.local_config) as db,
+        connect_database(cli_ctx.local_config.db) as db,
         db.begin_session() as session,
     ):
         alias_row = await session.scalar(
@@ -225,7 +225,7 @@ async def dealias(cli_ctx, alias):
 
 async def validate_image_alias(cli_ctx, alias: str) -> None:
     async with (
-        connect_database(cli_ctx.local_config) as db,
+        connect_database(cli_ctx.local_config.db) as db,
         db.begin_readonly_session() as session,
     ):
         try:
@@ -255,7 +255,7 @@ async def validate_image_canonical(
     cli_ctx, canonical: str, current: bool, architecture: Optional[str] = None
 ) -> None:
     async with (
-        connect_database(cli_ctx.local_config) as db,
+        connect_database(cli_ctx.local_config.db) as db,
         db.begin_readonly_session() as session,
     ):
         try:
