+++ src/ai/backend/manager/models/health.py
@@ -154,7 +154,7 @@ async def get_redis_object_info_list(root_ctx: RootContext) -> list[RedisObjectC
 
 
 async def _get_connnection_info(root_ctx: RootContext) -> ConnectionInfoOfProcess:
-    node_id = root_ctx.unified_config.local.manager.id or socket.gethostname()
+    node_id = root_ctx.unified_config.shared.manager.id or socket.gethostname()
     pid = os.getpid()
 
     sqlalchemy_info = await get_sqlalchemy_connection_info(root_ctx)
@@ -165,7 +165,7 @@ async def _get_connnection_info(root_ctx: RootContext) -> ConnectionInfoOfProces
 
 
 async def report_manager_status(root_ctx: RootContext) -> None:
-    lifetime = root_ctx.unified_config.local.manager.status_lifetime
+    lifetime = root_ctx.unified_config.shared.manager.status_lifetime
     cxn_info = await _get_connnection_info(root_ctx)
     _data = msgpack.packb(cxn_info.model_dump(mode="json"))
